make getCurrentUser controller and add GET /api/user/current to swagger
